ignore WIP files
